refactor: moved the 'utils' module to a more suitable location
