+++ api_inmobiliaria/apps/emails/infrastructure/views/account_management/send_token.py
@@ -1,8 +1,8 @@
 from apps.emails.infrastructure.repositories import TokenRepository
 from apps.emails.applications import ActionLinkManager
 from apps.users.infrastructure.repositories import UserRepository
-from apps.utils.generators import TokenGenerator
-from apps.utils.validators import is_valid_uuid
+from utils.generators import TokenGenerator
+from utils.validators import is_valid_uuid
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.generics import GenericAPIView
