+++ api_inmobiliaria/settings/environments/base.py
@@ -128,7 +128,7 @@ REST_FRAMEWORK = {
         "rest_framework.authentication.SessionAuthentication",
     ],
     "DEFAULT_PERMISSION_CLASSES": ["rest_framework.permissions.IsAuthenticated"],
-    "EXCEPTION_HANDLER": "apps.utils.exceptions.api_view_exception_handler",
+    "EXCEPTION_HANDLER": "utils.exceptions.api_view_exception_handler",
     "DEFAULT_SCHEMA_CLASS": "drf_spectacular.openapi.AutoSchema",
 }
