+++ api_inmobiliaria/tests/users/views/register/test_searcher.py
@@ -1,9 +1,5 @@
-from apps.users.constants import (
-    SearcherProperties,
-    BaseUserProperties,
-)
 from apps.api_exceptions import DatabaseConnectionAPIError
-from apps.utils.messages import ERROR_MESSAGES
+from utils.messages import ERROR_MESSAGES
 from tests.factory import UserFactory
 from tests.utils import fake
 from rest_framework import status
